Move returns schema code into its own service
