+++ app/services/data/tear-down/tear-down.service.js
@@ -8,13 +8,12 @@
 const { db } = require('../../../../db/db.js')
 const TearDownWaterSchemaService = require('./tear-down-water-schema.service.js')
 const TearDownCrmSchemaService = require('./tear-down-crm-schema.service.js')
+const TearDownReturnsSchemaService = require('./tear-down-returns-schema.service.js')
 
 async function go () {
   await TearDownWaterSchemaService.go()
   await TearDownCrmSchemaService.go()
-
-  // returns schema
-  await _deleteReturns()
+  await TearDownReturnsSchemaService.go()
 
   // permit schema
   await _deletePermit()
@@ -23,23 +22,6 @@ async function go () {
   await _deleteUsers()
 
   return 'Test data deleted!'
-}
-
-async function _deleteReturns () {
-  await db
-    .from('returns.lines as l')
-    .innerJoin('returns.versions as v', 'l.versionId', 'v.versionId')
-    .innerJoin('returns.returns as r', 'v.returnId', 'r.returnId')
-    .where('r.isTest', true)
-    .del()
-
-  await db
-    .from('returns.versions as v')
-    .innerJoin('returns.returns as r', 'v.returnId', 'r.returnId')
-    .where('r.isTest', true)
-    .del()
-
-  await _deleteTestData('returns.returns')
 }
 
 async function _deletePermit () {
@@ -54,13 +36,6 @@ async function _deleteUsers () {
     .from('idm.users')
     .where(db.raw("user_data->>'source' = 'acceptance-test-setup'"))
     .orWhereLike('userName', '%@example.com')
-    .del()
-}
-
-async function _deleteTestData (tableName) {
-  await db
-    .from(tableName)
-    .where('isTest', true)
     .del()
 }
